moves get_ban_turn_info to a method on Map
